% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSAproc.R
\name{boundary}
\alias{boundary}
\title{boundary}
\usage{
boundary(informationFractions, side, alpha, zninf = -8, tol = 1e-07)
}
\arguments{
\item{informationFractions}{Vector of information levels out of the required
information. Must be in cronological order and between 0 and 1.}

\item{side}{2 or 1, for 2-sided or 1-sided testing. For now is only 2-sided
possible}

\item{alpha}{alpha level e.g. 0.05}

\item{zninf}{First boundary on z-scale when the first alpha spending is too
small based on the variable tol. Defaults to -8.}

\item{tol}{The tolerance level on probabilities when defining the stopping
boundaries. Defaults to 1e-13.}
}
\value{
A list of two elements
\item{InformationFractions}{The information fractions. On a scale from 0 to 1.}
\item{alpha.boundaries}{The alpha spending boundaries. For 2-sided testing,
 the values are the upper boundaries. The lower boundaries are
 created by (-1)*alpha.boundaries}
}
\description{
Calculate boundaries for sequential meta-analysis. The functions are based
on the TSA software
}
\examples{
timing = c(0.19, 0.5, 1)
boundary(informationFractions = timing, side = 2, alpha = 0.05)

}
