% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coronary}
\alias{coronary}
\title{Dataset of studies investigating the intensity of statin therapy on the risk of myocardial infarction or coronary death}
\format{
A data frame with 4 rows and 5 variables:
\describe{
  \item{study}{Name of first author of the trial}
  \item{eI}{Number of events in the intervention group}
  \item{nI}{Number of pax in the intervention group}
  \item{eC}{Number of events in the control group}
  \item{nC}{Number of pax in the control group}
}
}
\usage{
coronary
}
\description{
A dataset containing data on myocardial infarction or coronary death among 
patients with acute coronary syndromes or chronic coronary artery disease 
of statin therapy intensity. The studies compared low intensities of statin to higher intensities.
}
\keyword{datasets}
