% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eds}
\alias{eds}
\title{A dataset containing data on ...}
\format{
A data frame with 9 studies and 8 variables:
\describe{
  \item{study}{Name of the city of the study}
  \item{year}{Year of the trial}
  \item{mI}{Mean ESD in intervention group}
  \item{mC}{Mean ESD in control group}
  \item{sdI}{SD in intervention group}
  \item{sdC}{SD in control group}
  \item{nI}{Number of participants in the intervention group}
  \item{nC}{Number of participants in the control group}
}
}
\usage{
eds
}
\description{
A dataset containing data on ...
}
\keyword{datasets}
