% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsa_classes.R
\docType{class}
\name{EOFstack-class}
\alias{EOFstack-class}
\title{Class "EOFstack"}
\description{
Result from \code{\link[rtsa]{rtsa.eof}} stored as a \code{\linkS4class{EOFstack}} object.
Slots for \code{EOFstack} objects include:
\itemize{
\item \emph{eof.modes}: the spatial representation of eof modes;
\item \emph{expansion_coefficients}: the eof expansion coefficients;
\item \emph{total_variance}: the dataset total variance;
\item \emph{explained_variance}: the explained variance by each eof mode;
\item \emph{center}: the raster of center values;
\item \emph{scale}: the raster of scale values.
}
}
\section{Objects from the class}{
 Objects are created by calls to \code{\link[rtsa]{rtsa.eof}}.
}

\seealso{
\code{\link[rtsa]{rtsa.eof}}
}
