% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsa_classes.R
\docType{class}
\name{EOFstack}
\alias{EOFstack}
\title{Class "EOFstack"}
\description{
Result from \code{\link[rtsa]{rtsa.eof}} stored as a \code{\linkS4class{EOFstack}} object.
Slots for \code{EOFstack} objects include: (1) the spatial representation of eof modes; 
(2) the eof expansion coefficients; (3) the dataset total variance;
(4) the explained variance from each eof; (5) the raster of center values; 
(6) the raster of scale values.
}
\section{Objects from the class}{
 Objects are created by calls to \code{\link[rtsa]{rtsa.eof}}.
}

\seealso{
\code{\link[rtsa]{rtsa.eof}}
}
