% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsa_classes.R
\docType{class}
\name{EOTstack-class}
\alias{EOTstack-class}
\title{Class "EOTstack"}
\description{
Result from \code{\link[rtsa]{rtsa.eot}} stored as a \code{\linkS4class{EOTstack}} object.
Slots for \code{EOTstack} objects include: 
\itemize{
\item \emph{eot}: the temporal profiles of the identified base point pixels;
\item \emph{total_variance}: the dataset total variance;
\item \emph{explained_variance}: the explained variance by each eot mode;
\item \emph{coords_bp}: the coordinates of the identified base point;
\item \emph{r_predictor}: the raster of correlation coefficients between the base point and each pixel of the predictor domain;
\item \emph{rsq_predictor}: the raster of coefficient of determination between the base point and each pixel of the predictor domain;
\item \emph{rsq_sums_predictor}: the raster of sums of coefficient of determination between the base point and each pixel of the predictor domain;
\item \emph{int_predictor}: the raster of intercept of the regression equation for each pixel of the predictor domain;
\item \emph{slp_predictor}: the raster of slope of the regression equation for each pixel of the predictor domain;
\item \emph{p_predictor}: the raster of significance (p-value) of the the regression equation for each pixel of the predictor domain.
}
}
\section{Objects from the class}{
 Objects are created by calls to \code{\link[rtsa]{rtsa.eot}}.
}

\seealso{
\code{\link[rtsa]{rtsa.eot}}
}
