% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaPrepare.R
\name{metaPrepare}
\alias{metaPrepare}
\title{metaPreapre}
\usage{
metaPrepare(
  outcome = "RR",
  eI = NULL,
  nI = NULL,
  eC = NULL,
  nC = NULL,
  mI = NULL,
  mC = NULL,
  sdI = NULL,
  sdC = NULL,
  vartype = "equal",
  method = "MH"
)
}
\arguments{
\item{outcome}{binary outcome of interest (option include: RD, OR, RR)}

\item{eI}{number of events in intervention group}

\item{nI}{number of pax in intervention group}

\item{eC}{number of events in control group}

\item{nC}{number of pax in control group}

\item{mI}{effect estimate in intervention group (continuous outcome)}

\item{mC}{effect estimate in control group (continuous outcome)}

\item{sdI}{standard error of estimate in intervention group (continuous outcome)}

\item{sdC}{standard error of estimate in control group (continuous outcome)}

\item{vartype}{Variance type for continuous outcomes. Default is "equal", other choice is "non-equal"}

\item{method}{Method for calculating weights. Options include: MH (Mantel-Haenzel), Inverse variance weighting (IV) or GLM}
}
\value{
A list is returned with the following items:
\item{w}{weights}
\item{te}{trial specific effect size}
\item{lower}{lower CI for trial specific effect size}
\item{upper}{upper CI for trial specific effect size}
\item{pe}{pooled effect size (for MH method svpe is also included which is the sum of variance)}
\item{sig}{sig: sigma (standard deviation of effect size per trial - consider renaming)}
\item{outcome}{statistic used for effect size (OR, RR og RD)}
\item{method}{method for calculating weights. GLM, Inverse variance weighting (IV) or Mantel-Haenzel (MH)}
\item{eI}{number of events in the intervention group}
\item{eC}{number of events in the control group}
\item{nI}{number of pax in the intervention group}
\item{nC}{number of pax in the control group}
}
\description{
Preparation for synthesization of results for binary data. Calculates study specific effect sizes, standard errors, confidence intervals and weights to use for meta-analysis. Handling of continuous outcomes will be added soon.
}
\examples{
data(perioOxy)
with(perioOxy, metaPrepare(outcome = "RR", eI = eI, nI = nI, eC = eC, nC = nC,
method = "MH"))
}
