% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minTrial.R
\name{minTrial}
\alias{minTrial}
\title{minTrial}
\usage{
minTrial(
  metric,
  value,
  tau2,
  alpha = 0.05,
  beta = 0.2,
  p0 = NULL,
  verbose = FALSE
)
}
\arguments{
\item{metric}{Metric of interest, options include only risk ratio (RR) for now}

\item{value}{Value of minimum relevant difference}

\item{tau2}{Heterogeneity estimate}

\item{alpha}{Type 1 error}

\item{beta}{Type 2 error}

\item{p0}{If verbose i set to TRUE and metric is set to RR (risk ration),
define common probability of event p0}

\item{verbose}{TRUE or FALSE. If the number of required participants per
trial is wanted as well, a matrix with the number of trials and the number
of required participants per trial is printed.}
}
\value{
Either a number (minimum required trials) or the minimum required
required trials together with a matrix of required participants per trial given
different number of trials.
}
\description{
Calculate minimum number of trials for wanted power in a meta-analysis with
heterogeneity
}
\examples{
minTrial(metric = "RR", value = 0.7, tau2 = 0.05)

}
