% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsa_classes.R
\docType{class}
\name{MKstack-class}
\alias{MKstack-class}
\title{Class "MKstack"}
\description{
Result from \code{\link[rtsa]{rtsa.mk}} stored as a \code{\linkS4class{MKstack}} object.
Slots for \code{MKstack} objects include: 
\itemize{
\item \emph{tau}: the raster of Kendall tau statistic;
\item \emph{pvalue}: the raster of Kendall two-sided p-value;
\item \emph{score}: the raster of Kendall Score;
\item \emph{variance}: the raster of Variance of Kendall Score.
}
}
\section{Objects from the class}{
 Objects are created by calls to \code{\link[rtsa]{rtsa.mk}}.
}

\seealso{
\code{\link[rtsa]{rtsa.stl}}, \code{\link[rtsa]{rtsa.seas}}, \code{\link[rtsa]{rtsa.mk}}
}
