% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nFixed.R
\name{nFixed}
\alias{nFixed}
\title{nFixed}
\usage{
nFixed(alpha, beta, pI = NULL, pC = NULL, iE = NULL, sdE = NULL, binary = TRUE)
}
\arguments{
\item{alpha}{type 1 error}

\item{beta}{type 2 error}

\item{pI}{probability of event in intervention group (for binary outcomes)}

\item{pC}{probability of event in control group (for binary outcomes)}

\item{iE}{intervention effect (for continuous outcomes)}

\item{sdE}{standard deviation of intervention effect (for continuous outcomes)}

\item{binary}{TRUE/FALSE. If continuous outcome set to FALSE. Default is TRUE.}
}
\value{
The total required number of  participants to achieve a specific power
in a fixed-effect model
}
\description{
Calculate the number of total participants when there is no presence
 of heterogeneity to achieve a specific power.
}
\examples{
log.RR = log(0.9)
p0 = 0.1
pI = exp(log(p0)+log.RR/2)
pC = exp(log(p0)-log.RR/2)
nFixed(alpha = 0.05, beta = 0.2, pI = pI, pC = pC, binary = TRUE)

}
