% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nRandom.R
\name{nRandom}
\alias{nRandom}
\title{nRandom}
\usage{
nRandom(alpha, beta, pI, pC, diversity = NULL)
}
\arguments{
\item{alpha}{type 1 error}

\item{beta}{type 2 error}

\item{pI}{proportion of events in intervention group}

\item{pC}{proportion of events in control group}

\item{diversity}{estimate of diversity. Must be between 0 and 1.}
}
\value{
RIS for random-effects meta-analysis
}
\description{
Calculate the RIS using diversity. Implementation of method from TSA software.
}
\examples{
log.RR = log(0.9)
p0 = 0.1
pI = exp(log(p0)+log.RR/2)
pC = exp(log(p0)-log.RR/2)
nRandom(alpha = 0.05, beta = 0.2, pI = pI, pC = pC, diversity = 0.7)

}
