% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patchcea}
\alias{patchcea}
\title{A dataset containing data on ...}
\format{
A data frame with 6 rows and 5 variables:
\describe{
  \item{study}{Name of first author of the trial}
  \item{eI}{Number of events in the intervention group}
  \item{nI}{Number of pax in the intervention group}
  \item{eC}{Number of events in the control group}
  \item{nC}{Number of pax in the control group}
}
}
\usage{
patchcea
}
\description{
A dataset containing data on ...
}
\keyword{datasets}
