% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perioOxy}
\alias{perioOxy}
\title{Dataset of RCTs investigating the effect of 80\% perioperative oxygen vs. 30
-35\% perioperative oxygen on surgical site infection.}
\format{
A data frame with 7 rows and 6 variables:
\describe{
  \item{study}{Name of first author of the trial}
  \item{year}{Year of trial}
  \item{eI}{Number of events in the intervention group (80\% oxygen)}
  \item{nI}{Number of pax in the intervention group (80\% oxygen)}
  \item{eC}{Number of events in the control group (30-35\% oxygen)}
  \item{nC}{Number of pax in the control group (30-35\% oxygen)}
}
}
\usage{
perioOxy
}
\description{
A dataset containing data on seven trials which includes their number of events per treatment group, where intervention is 80\% oxygen and control is 30-35\% oxygen, number of participants in each treatment group and the year of the trial.
}
\keyword{datasets}
