% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTSA_plot.R
\name{plot.RTSA}
\alias{plot.RTSA}
\title{plot.RTSA}
\usage{
\method{plot}{RTSA}(x, ...)
}
\arguments{
\item{x}{RTSA object}

\item{...}{Other arguments to plot.RTSA}
}
\value{
Plot. Either a plot for two sided testing or one-sided
}
\description{
plot.RTSA
}
\examples{
data(perioOxy)
outRTSA <- RTSA(data = perioOxy, outcome = "RR", mc = 0.7, side = 2, alpha = 0.05, fixed = FALSE)
plot(x = outRTSA)

}
