% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsa.eot.R
\name{rtsa.eot}
\alias{rtsa.eot}
\title{EOT (Empirical Orthogonal Teleconnections analysis)}
\usage{
rtsa.eot(rasterts, rastermask = NULL, nu = NULL, gapfill = "none",
  predictor = "all", standardised = TRUE, reduce.both = FALSE,
  type = "rsq", verbose = FALSE)
}
\arguments{
\item{rasterts}{Input raster time series as \code{\linkS4class{RasterStackTS}} or \code{\linkS4class{RasterBrickTS}} object.}

\item{rastermask}{Either a \code{\linkS4class{RasterLayer}} or "compute". Raster layer to use as a mask. When "compute" 
is set raster mask is computed to remove all pixels with incomplete time series.}

\item{nu}{Numeric. Defines the number of EOTs to return. Defaults to return the first 2 EOT modes.}

\item{gapfill}{Character. Defines the algorithm to be used to interpolate pixels with incomplete temporal profiles
Accepts argument supported as method in function \code{\link[rtsa]{rtsa.gapfill}}.}

\item{predictor}{Character. Defines the predictor components to export from those available in Value section of \code{\link[remote]{eot}}.
Supports one or more of the following: 'all', 'r_predictor', 'rsq_predictor', 'rsq_sums_predictor', 'int_predictor', 'slp_predictor', 'p_predictor'.}

\item{...}{Additional arguments to be passed through to function \code{\link[remote]{eot}}.}
}
\value{
Object of class \code{\link{EOTstack-class}} containing the following components:
\tabular{rll}{
\tab \code{eot} \tab EOT temporal profiles corresponding to base point coordinates as \code{\linkS4class{xts}} object\cr
\tab \code{total_variance} \tab Numeric. Total explained variance of input raster time series by the entire set of computed EOTs\cr
\tab \code{explained_variance} \tab Numeric vector. Percentage of variance explained by each EOT mode with respect to the total variance of input raster time series\cr
\tab \code{r_predictor} \tab RasterBrick. Correlation coefficients between the base point and each pixel of the predictor domain as \code{\linkS4class{RasterBrick}} object (only exported if \code{predictor = "all"} or \code{predictor = "r_predictor"})\cr
\tab \code{rsq_predictor} \tab RasterBrick. Coefficient of determination between the base point and each pixel of the predictor domain as \code{\linkS4class{RasterBrick}} object (only exported if \code{predictor = "all"} or \code{predictor = "rsq_predictor"})\cr
\tab \code{rsq_sums_predictor} \tab RasterBrick. Sums of correlation coefficients between the base point and each pixel of the predictor domain as \code{\linkS4class{RasterBrick}} object (only exported if \code{predictor = "all"} or \code{predictor = "rsq_sums_predictor"})\cr
\tab \code{int_predictor} \tab RasterBrick. Intercept of the regression equation for each pixel of the predictor domain as \code{\linkS4class{RasterBrick}} object (only exported if \code{predictor = "all"} or \code{predictor = "int_predictor"})\cr
\tab \code{slp_predictor} \tab RasterBrick. Slope of the regression equation for each pixel of the predictor domain as \code{\linkS4class{RasterBrick}} object (only exported if \code{predictor = "all"} or \code{predictor = "slp_predictor"})\cr
\tab \code{r_predictor} \tab RasterBrick. The significance (p-value) of the the regression equation for each pixel of the predictor domain as \code{\linkS4class{RasterBrick}} object (only exported if \code{predictor = "all"} or \code{predictor = "p_predictor"})\cr
}
}
\description{
This function conducts an Empirical Orthogonal Teleconnection analysis (EOT) of a raster time series 
predictor using "remote" package especially designed to identify locations of enhanced potential and explain 
spatio-temporal variability within the same geographic field.
}
\details{

}
\examples{
\dontrun{
## create raster time series using the 'pacificSST' data from 'remote' package
require(remote)

data(pacificSST)
pacificSST[which(getValues(pacificSST == 0))] <- NA # set NA values
# create rts object
rasterts <- rts(pacificSST, seq(as.Date('1982-01-15'), as.Date('2010-12-15'), 'months'))

## generate raster mask
rmk <- pacificSST[[1]] # create raster mask
values(rmk) <- 1 # set raster mask values
rmk[which(is.na(getValues(pacificSST[[1]])))] <- 0 # set raster mask values

## compute EOT
# compute the first 2 EOTs
eot_result <- rtsa.eot(rasterts=rasterts, rastermask=rmk, nu=2)
# compute the first 2 EOTs and export only the compontents 'r_predict' and 'p_predictor'
eot_rp <- rtsa.eot(rasterts=rasterts, rastermask="compute", nu=2, predictor=c("r_predictor", "p_predictor"))
# compute the first 2 EOTs using the index of agreement
eot_ioa <- rtsa.eot(rasterts=rasterts, rastermask=rmk, nu=2, type="ioa", predictor="all", verbose=T)
}

}
\references{
van den Dool H.M., Saha S., Johansson A. (2000). Empirical Orthogonal Teleconnections. 
Journal of Climate, Volume 13, Issue 8, pp. 1421-1435. 
\url{http://journals.ametsoc.org/doi/abs/10.1175/1520-0442%282000%29013%3C1421%3AEOT%3E2.0.CO%3B2
}
}
\seealso{
\code{\link[remote]{eot}}, \code{\link[rtsa]{rtsa.eof}}, \code{\link[rtsa]{rtsa.gapfill}}, \code{\linkS4class{EOTstack}}
}
\author{
Federico Filipponi
}
\keyword{EOT}
\keyword{analysis}
\keyword{series}
\keyword{time}
