% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsa.mk.R
\name{rtsa.mk}
\alias{rtsa.mk}
\title{Raster time series Mann-Kendall trend test}
\usage{
rtsa.mk(rasterts, rastermask = NULL, gapfill = "none", cores = 1L,
  verbose = FALSE)
}
\arguments{
\item{rasterts}{Input raster time series as \code{\linkS4class{RasterStackTS}} or \code{\linkS4class{RasterBrickTS}} object.}

\item{rastermask}{Either a \code{\linkS4class{RasterLayer}} or "compute". Raster layer to use as a mask. When "compute" 
is set raster mask is computed to remove all pixels with incomplete time series.}

\item{gapfill}{Character. Defines the algorithm to be used to interpolate pixels with incomplete temporal profiles. 
Accepts argument supported as method in function \code{\link[rtsa]{rtsa.gapfill}}.}

\item{cores}{Integer. Defines the number of CPU to be used for multicore processing. Default to "1" core for 
singlecore processing.}

\item{...}{Additional arguments to be passed through to function \code{\link[Kendall]{MannKendall}}.}
}
\value{
Object of class \code{\link{MKstack-class}} containing the following components:
\tabular{rll}{
\tab \code{tau} \tab Kendall tau statistic\cr
\tab \code{pvalue} \tab Kendall two-sided p-value\cr
\tab \code{score} \tab Kendall Score\cr
\tab \code{variance} \tab Variance of Kendall Score\cr
}
}
\description{
This function conducts Mann-Kendall trend test from raster time series
using "Kendall" package especially designed to handle gappy time series.
}
\details{

}
\examples{
\dontrun{
## create raster time series using the 'pacificSST' data from 'remote' package
require(remote)

data(pacificSST)
pacificSST[which(getValues(pacificSST == 0))] <- NA # set NA values
# create rts object
rasterts <- rts(pacificSST, seq(as.Date('1982-01-15'), as.Date('2010-12-15'), 'months'))

## generate raster mask
raster_mask <- pacificSST[[1]] # create raster mask
names(raster_mask) <- "mask"
values(raster_mask) <- 1 # set raster mask values
raster_mask[which(is.na(getValues(pacificSST[[1]])))] <- 0 # set raster mask values

# compute Mann-Kendall trend test
MannKendall_result <- rtsa.mk(rasterts=rasterts, rastermask=raster_mask)
# compute Mann-Kendall trend test using multiple cores on monthly time series
### create monthly averages
rasterts_monthly_mean <- apply.monthly(rasterts, mean)
MannKendall_monhtly_result <- rtsa.mk(rasterts=rasterts_monthly_mean, rastermask=raster_mask)
}

}
\references{
Mann, H.B. (1945). Non-parametric tests against trend. Econometrica, 13, 163-171.
Kendall, M.G. (1975). Rank Correlation Methods, 4th edition. Charles Griffin, London.
Gilbert, R.O. (1987) . Statistical Methods for Environmental Pollution Monitoring. Wiley, NY.
Davison, A.C. and Hinkley, D.V. (1997) Bootstrap Methods and Their Application. Cambridge University Press.
Hipel, K.W. and McLeod, A.I., (2005). Time Series Modelling of Water Resources and Environmental Systems. 
Electronic reprint of our book orginally published in 1994. \href{http://www.stats.uwo.ca/faculty/aim/1994Book/}{book}
}
\seealso{
\code{\link[Kendall]{MannKendall}}, \code{\link[rtsa]{rtsa.stl}}, \code{\link[rtsa]{rtsa.seas}}, \code{\link[rtsa]{rtsa.gapfill}}
}
\author{
Federico Filipponi
}
\keyword{Mann-Kendall}
\keyword{analysis}
\keyword{series}
\keyword{test}
\keyword{time}
\keyword{trend}
