% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsa.scaleEOF.R
\name{rtsa.scaleEOF}
\alias{rtsa.scaleEOF}
\title{Scale EOF modes}
\usage{
rtsa.scaleEOF(x, cut = 1, method = "percentile")
}
\arguments{
\item{x}{Input \code{\link{EOFstack-class}} object generated using \code{\link[rtsa]{rtsa.eof}}}

\item{cut}{Numeric. Defines the percentile to be used to scale each EOFs. Default is set to 1 (1st and 99th percentiles)}

\item{method}{Character. Method used to scale EOF modes. Only default \code{"percentile"} is supported}
}
\value{
Updates the 'eof' slot in the object of class \code{\link{EOFstack-class}} containing the following components:
\tabular{lll}{
\tab \code{eof} \tab Scaled EOF modes as \code{\linkS4class{RasterBrick}} object\cr
\tab \code{expansion_coefficients} \tab EOF Expansion Coefficients (EC) as \code{\linkS4class{xts}} object\cr
\tab \code{total_variance} \tab Numeric. Total variance of input raster time series\cr
\tab \code{explained_variance} \tab Numeric vector. Percentage of variance explained by each EOF mode with respect to the total variance of input raster time series\cr
\tab \code{center} \tab Center values from each pixel temporal profile as \code{\linkS4class{RasterLayer}} object (only computed if \code{centered = TRUE})\cr
\tab \code{scale} \tab Scale values from each pixel temporal profile as \code{\linkS4class{RasterLayer}} object (only computed if \code{scaled = TRUE})
}
}
\description{
This function rescale resulting spatial modes generated from Empirical Orthogonal Function analysis (EOF)
in the range between -1 and 1
}
\details{

}
\examples{
\dontrun{
## Scale EOF modes by clipping the distribution between 1st and 99th percentile
scaled_eof <- rtsa.scaleEOF(x=eof_result, cut=1)
## Scale EOF modes by clipping the distribution between 5th and 95th percentile
scaled_eof <- rtsa.scaleEOF(x=eof_result, cut=5)
}

}
\seealso{
\code{\link[rtsa]{rtsa.eof}}, \code{\link[sinkr]{eof}}
}
\author{
Federico Filipponi
}
\keyword{EOF}
\keyword{PCA}
\keyword{SVD}
\keyword{analysis}
\keyword{series}
\keyword{time}
