% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsa.stl.R
\name{rtsa.stl}
\alias{rtsa.stl}
\title{Raster time series STL (Seasonal Trend Decomposition using Loess)}
\usage{
rtsa.stl(rasterts, rastermask = NULL, gapfill = "none", cores = 1L,
  n.p = NULL, s.window = "periodic", t.window = NULL, s.degree = 1L,
  t.degree = 1L, only.statistics = FALSE, keep.original = FALSE,
  verbose = FALSE)
}
\arguments{
\item{rasterts}{Input raster time series as \code{\linkS4class{RasterStackTS}} or \code{\linkS4class{RasterBrickTS}} object.}

\item{rastermask}{Either a \code{\linkS4class{RasterLayer}} or "compute". Raster layer to use as a mask. When "compute" 
is set raster mask is computed to remove all pixels with incomplete time series.}

\item{gapfill}{Character. Defines the algorithm to be used to interpolate pixels with incomplete temporal profiles. 
Accepts argument supported as method in function \code{\link[rtsa]{rtsa.gapfill}}.}

\item{cores}{Integer. Defines the number of CPU to be used for multicore processing. Default to "1" core for 
singlecore processing.}

\item{n.p}{Integer. Argument to be passed to function \code{\link[stlplus]{stlplus}}. Periodicity of the seasonal component. 
Default to the frequency estimated from of the time series.}

\item{s.window}{Argument to be passed to function \code{\link[stlplus]{stlplus}}. Either the character string "periodic" (default) 
or the span (in lags) of the loess window for seasonal extraction, which should be odd.}

\item{t.window}{Integer. Argument to be passed to function \code{\link[stlplus]{stlplus}}. The span (in lags) of the the loess window for trend extraction, 
which should be odd. Default to the entire time series duration.}

\item{s.degree}{Integer. Argument to be passed to function \code{\link[stlplus]{stlplus}}. Degree of locally-fitted polynomial in trend extraction.
Should be 0, 1 (default) or 2.}

\item{t.degree}{Integer. Argument to be passed to function \code{\link[stlplus]{stlplus}}. Degree of locally-fitted polynomial in trend extraction.
Should be 0, 1 (default) or 2.}

\item{only.statistics}{Logical. If TRUE returns only the statistics from seasonal, trend and remainder components.}

\item{keep.original}{Logical. If TRUE returns the original raster time series values in the 'rts' slot of \code{\link{STDstack-class}} object.}

\item{...}{Additional arguments to be passed through to function \code{\link[stlplus]{stlplus}}.}
}
\value{
Object of class \code{\link{STDstack-class}} containing the following components:
\tabular{rll}{
\tab \code{std} \tab Seasonal Trend Decomposition method used\cr
\tab \code{mask} \tab Final raster mask of computed pixels as \code{\linkS4class{RasterLayer}} object\cr
\tab \code{seasonal_amplitude} \tab Amplitude of seasonal component (statistic) as \code{\linkS4class{RasterLayer}} object\cr
\tab \code{seasonal_amplitude_stdev} \tab Standard deviation computed from the amplitude of seasonal component (statistic) as \code{\linkS4class{RasterLayer}} object (only returned when running function \code{\link[rtsa]{rtsa.seas}})\cr
\tab \code{trend_slope} \tab Trend slope computed from trend component (yearly statistic) as \code{\linkS4class{RasterLayer}} object\cr
\tab \code{residual_stdev} \tab Standard deviation computed from the remainder component (statistics) as \code{\linkS4class{RasterLayer}} object\cr
\tab \code{rts} \tab Input raster time series as \code{\linkS4class{RasterBrickTS}} object (only returned if \code{keep.original = TRUE})\cr
\tab \code{seasonality} \tab Seasonal component as \code{\linkS4class{RasterBrickTS}} object\cr
\tab \code{trend} \tab Trend component as \code{\linkS4class{RasterBrickTS}} object\cr
\tab \code{seasonaladjtrend} \tab Seasonal adjusted trend component as \code{\linkS4class{RasterBrickTS}} object (only returned when running function \code{\link[rtsa]{rtsa.seas}})\cr
\tab \code{remainder} \tab Remainder component as \code{\linkS4class{RasterBrickTS}} object\cr
}
}
\description{
This function conducts a Seasonal Trend Decomposition using Loess (STL) from raster time series
using "stlplus" package especially designed to handle gappy time series.
}
\details{

}
\examples{
\dontrun{
## create raster time series using the 'pacificSST' data from 'remote' package
require(remote)

data(pacificSST)
pacificSST[which(getValues(pacificSST == 0))] <- NA # set NA values
# create rts object
rasterts <- rts(pacificSST, seq(as.Date('1982-01-15'), as.Date('2010-12-15'), 'months'))

## generate raster mask
raster_mask <- pacificSST[[1]] # create raster mask
names(raster_mask) <- "mask"
values(raster_mask) <- 1 # set raster mask values
raster_mask[which(is.na(getValues(pacificSST[[1]])))] <- 0 # set raster mask values

## compute Seasonal Trend Decomposition analysis
# compute 'STL'
std_result <- rtsa.stl(rasterts=rasterts, rastermask=raster_mask)
# compute STL' with multiple cores support and returning the original raster values
std_result <- rtsa.stl(rasterts=rasterts, rastermask=raster_mask, cores=2, keep.original=TRUE)
# compute STL' using additional arguments to define loess window
std_result <- rtsa.stl(rasterts=rasterts, rastermask=raster_mask, n.p=12, t.window=48)
}

}
\references{
Cleveland, R.B., Cleveland, W.S., Terpenning, I. (1990). STL: A seasonal-trend decomposition procedure based on loess. 
Journal of Official Statistics, 6(1), 3.

Hafen, R.P. (2010). Local regression models: Advancements, applications, and new methods. 
West Lafayette, Indiana: Purdue University, PhD dissertation, pp. 279.
}
\seealso{
\code{\link[stlplus]{stlplus}}, \code{\link[rtsa]{rtsa.seas}}, \code{\link[rtsa]{rtsa.gapfill}}, \code{\link[seasonal]{seas}}, \code{\link[stats]{stl}}, \code{\link[stats]{decompose}}
}
\author{
Federico Filipponi
}
\keyword{Loess}
\keyword{STL}
\keyword{Seasonal}
\keyword{Trend}
\keyword{analysis}
\keyword{decomposition}
\keyword{series}
\keyword{time}
