% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsa_classes.R
\docType{class}
\name{STDstack-class}
\alias{STDstack-class}
\title{Class "STDstack"}
\description{
Result from \code{\link[rtsa]{rtsa.stl}} and \code{\link[rtsa]{rtsa.seas}} stored as a \code{\linkS4class{STDstack}} object.
Slots for \code{STDstack} objects include: 
\itemize{
\item \emph{std}: Seasonal Trend Decomposition method used;
\item \emph{mask}: the raster mask of valid pixels;
\item \emph{seasonal_amplitude}: the raster of seasonal amplitude;
\item \emph{seasonal_amplitude}: the raster of seasonal amplitude standard deviation;
\item \emph{trend_slope}: the raster of trend slope value;
\item \emph{remainder_stdev}: the raster of the remainder standard deviation;
\item \emph{rts}: input raster time series as RasterBrickTS object;
\item \emph{seasonality}: seasonal component as RasterBrickTS object;
\item \emph{trend}: trend component as RasterBrickTS object;
\item \emph{seasonaladjtrend}: seasonal adjusted trend component as RasterBrickTS object;
\item \emph{remainder}: remainder component as RasterBrickTS object.
}
}
\section{Objects from the class}{
 Objects are created by calls to \code{\link[rtsa]{rtsa.stl}} and \code{\link[rtsa]{rtsa.seas}}.
}

\seealso{
\code{\link[rtsa]{rtsa.stl}}, \code{\link[rtsa]{rtsa.seas}}
}
