% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize.R
\name{synthesize}
\alias{synthesize}
\title{synthesize}
\usage{
synthesize(y, sign = NULL, random = TRUE, fixedStudy = TRUE, hakn = FALSE)
}
\arguments{
\item{y}{list-object. Preferably a "synthPrepped" object from the metaPrepare function.}

\item{sign}{TBA}

\item{random}{TRUE or FALSE. Whether a random-effects model estimate and belonging statistics should be calculated. Defaults to TRUE.}

\item{fixedStudy}{TRUE or FALSE. When using GLM methods for the meta-analysis, should the study effect be fixed-effect. Defaults to TRUE.}

\item{hakn}{TRUE or FALSE. When conducting a random-effects model should the Harting-Knapp-Sidik-Jonkman adjustment be used. Else is DerSimonian-Laird used. Defaults to FALSE.}
}
\value{
A list is returned with the a set of the following items:
\item{fw}{Weights for the fixed-effect model}
\item{peF}{Vector containing fixed-effect model results - pooled effect size, lower 95\% confidence limit, upper 95\% confidence limit, z-value, p-value, log pooled effect size, variance of estimate}
\item{rwR}{Weights for the random-effects model}
\item{peR}{Vector containing random-effects model results - pooled effect size, lower 95\% confidence limit, upper 95\% confidence limit, z-value (DL) or t-value (HKSJ), p-value, variance of estimate}
\item{Q}{vector containing Q-measure, degrees of freedom (df) and p-value for Q}
\item{}{vector containing tau2 (random effect variance estimate), H-measure, I2 (inconsistency) and D2 (diversity)}
}
\description{
Synthesize trials in a meta-analysis
}
\examples{
data(perioOxy)
m.prepare = with(perioOxy, metaPrepare(outcome = "RR", eI = eI,
 nI = nI, eC = eC, nC = nC, method = "MH"))
synthesize(m.prepare)

}
