% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eds}
\alias{eds}
\title{Dataset of studies investigating the effect of carer on early supported discharge services}
\format{
A data frame with 9 studies and 8 variables:
}
\usage{
eds
}
\description{
A dataset containing data on the length of hospital stay when receiving early
supported discharge (ESD) service versus conventional care. The outcome is
length of initial hospital stay counted in days.
}
\details{
\itemize{
  \item study. Name of the city of the study
  \item year. Year of the trial
  \item mI. Mean duration at hospital in intervention (ESD) group
  \item mC. Mean duration at hospital in control group
  \item sdI. Standard deviation of intervention (ESD) estimate
  \item sdC. Standard deviation of control estimate
  \item nI. Number of participants in the intervention (ESD) group
  \item nC. Number of participants in the control group
}
}
\references{
Fearon P, Langhorne P. Services for reducing duration of hospital care for acute stroke patients. Cochrane Database of Systematic Reviews 2012, Issue 9. Art. No.: CD000443. DOI: 10.1002/14651858.CD000443.pub3. Accessed 17 October 2022.
}
\keyword{datasets}
