% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaanalysis.R
\name{metaanalysis}
\alias{metaanalysis}
\alias{print.metaanalysis}
\title{metaanalysis}
\usage{
metaanalysis(
  data = NULL,
  outcome = "RR",
  vartype = "unequal",
  method = "MH",
  fixedStudy = TRUE,
  hksj = FALSE,
  sign = NULL,
  ...
)
}
\arguments{
\item{data}{Data.frame. If data is provided as a data set. Dataset must then containing arguments for
meta-analysis. Either `study`, `eI`, `eC`, `nC` or `nI` for discrete data, or, `study`, `mI`, `mC`, `sdI` and `sdC` for cont.
See details. If a data frame is not provided, the columns must be set in the function.}

\item{outcome}{Outcome metric for the studies. Choose between: cont, RR, RD or OR}

\item{vartype}{Variance type for continuous outcomes. Choices are "equal" or "non-equal". Defaults to "equal".}

\item{method}{Method for calculating weights. Options include: MH (Mantel-Haenzel), Inverse variance weighting (IV)
or GLM}

\item{fixedStudy}{TRUE or FALSE. For using GLM methods for the meta-analysis, should the study effect be
fixed-effect. Defaults to TRUE.}

\item{hksj}{TRUE or FALSE. Should the Hartung-Knapp-Sidik-Jonkman adjustment be used to the random-effects
meta-analysis. Defaults to FALSE.}

\item{sign}{TBA.}

\item{...}{Additional variables. See Details.}
}
\value{
A list object containing two data frames.
1. studyResults contains information about the individual studies
2. metaResults contains information about the meta-analysis (traditional, non-sequential)
}
\description{
metaanalysis
}
\examples{
data(perioOxy)
metaanalysis(outcome = "RR", data = perioOxy, study = perioOxy$trial)
data(eds)
metaanalysis(outcome = "cont", data = eds, study = eds$study)
}
