% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minTrial.R
\name{minTrial}
\alias{minTrial}
\title{minTrial}
\usage{
minTrial(
  outcome,
  delta,
  tau2,
  alpha = 0.05,
  beta = 0.2,
  p0 = NULL,
  var_delta = NULL,
  var_random = NULL
)
}
\arguments{
\item{outcome}{Metric of interest, options include only risk ratio (RR) for now}

\item{delta}{Value of minimum relevant difference}

\item{tau2}{Heterogeneity estimate}

\item{alpha}{Type 1 error}

\item{beta}{Type 2 error}

\item{p0}{If verbose i set to TRUE and outcome is set to RR (risk ration),
define common probability of event p0}

\item{var_delta}{Variance of the estimated effect}

\item{var_random}{Estimated variance from the random-effects meta-analysis}
}
\value{
Either a number (minimum required trials) or the minimum required
required trials together with a matrix of required participants per trial given
different number of trials.
}
\description{
Calculate minimum number of trials for wanted power in a meta-analysis with
heterogeneity
}
\examples{
minTrial(outcome = "RR", delta = 0.7, tau2 = 0.05)

}
