% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{plot.boundaries}
\alias{plot.boundaries}
\title{plot.boundaries}
\usage{
\method{plot}{boundaries}(x, theme = "classic", ...)
}
\arguments{
\item{x}{boundaries object}

\item{theme}{Whether the theme is "classic" or "aussie"}

\item{...}{Other arguments to plot.boundaries}
}
\value{
Plot. Either a plot for two- or one-sided testing.
}
\description{
plot.boundaries
}
\examples{
bounds <- boundaries(timing = c(0.5,0.75, 1), alpha  = 0.025, beta = 0.2,
side = 1, futility = "none", es_alpha = "esOF")
plot(x = bounds)

}
