% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTSA.R
\name{plot.RTSA}
\alias{plot.RTSA}
\title{plot.RTSA}
\usage{
\method{plot}{RTSA}(x, ylim = c(-8, 8), ...)
}
\arguments{
\item{x}{RTSA object}

\item{ylim}{y axis limits}

\item{...}{Other arguments to plot.RTSA}
}
\value{
Plot. Either a plot for two sided testing or one-sided
}
\description{
plot.RTSA
}
\examples{
data(perioOxy)
RTSAout = RTSA(data = perioOxy, outcome = "RR", mc = 0.9)
plot(x = RTSAout)

}
