% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTSA.R
\name{plot.RTSA}
\alias{plot.RTSA}
\title{plot.RTSA}
\usage{
\method{plot}{RTSA}(x, ...)
}
\arguments{
\item{x}{RTSA object}

\item{...}{Other arguments to plot.RTSA}
}
\value{
Plot. Either a plot for two sided testing or one-sided
}
\description{
plot.RTSA
}
\examples{
data(perioOxy)
count <- cumsum(perioOxy$nI+perioOxy$nC)
timing <- c(count/5540,1)
mp = metaPrepare(outcome = "RR", eI = perioOxy$eI, nI = perioOxy$nI,
eC = perioOxy$eC, nC = perioOxy$nC, method = "IV")
RTSAout = RTSA(timing = timing, synth = mp, anaTimes = c(4,5,7,8),
side = 2, alpha = 0.05)
plot(x = RTSAout)

}
