% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ris.R
\name{ris}
\alias{ris}
\alias{print.ris}
\title{ris: Calculate required information size and required number of trials.}
\usage{
ris(
  outcome,
  delta,
  side = 2,
  alpha = 0.05,
  beta = 0.2,
  random = FALSE,
  sd_delta = NULL,
  p0 = NULL,
  data = NULL,
  tau2 = NULL,
  I2 = NULL,
  D2 = NULL,
  var_random = NULL,
  type = "prospective",
  ...
)
}
\arguments{
\item{outcome}{Choose between: "cont", "RR", "OR" or "RD".}

\item{delta}{Minimum clinical relevant effect. For "OR" or "RR" set to natural scale, not log scale.}

\item{side}{Test type. Set to 1 or 2 depending on the test being 1- or 2-sided.}

\item{alpha}{Level of type-I-error. Default value is 0.05.}

\item{beta}{Level of type-II-error. Default value is 0.2.}

\item{random}{Should sample size be based on a random effects model? Defaults to FALSE.}

\item{sd_delta}{Standard deviation of estimated effect. Only needed when outcome = "cont".}

\item{p0}{Probability of event in population.}

\item{data}{Data set containing studies used for meta-analysis.}

\item{tau2}{The value of the heterogeneity. Use when estimating the sample size under a random effects model. If data is provided, the estimated heterogeneity is used instead.}

\item{I2}{Inconsistency.}

\item{D2}{Diversity.}

\item{var_random}{Estimated variance from the random-effects meta-analysis}

\item{type}{Prospective or retrospective sample size calculation.}

\item{...}{additional arguments}
}
\value{
A list containing required information size and more.
}
\description{
RTSA's sample size calculation.
}
\examples{
data("perioOxy")
ris(outcome = "RR", delta = 0.7, p0 = 0.1, data = perioOxy, random = FALSE)
}
