% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsa.eof.R
\name{rtsa.eof}
\alias{rtsa.eof}
\title{EOF (Empirical Orthogonal Functions analysis)}
\usage{
rtsa.eof(rasterts, rastermask = NULL, nu = NULL, gapfill = "none",
  centered = TRUE, scaled = FALSE, method = "svds", recursive = FALSE,
  cores = 1L, verbose = FALSE)
}
\arguments{
\item{rasterts}{Input raster time series as \code{\linkS4class{RasterStackTS}} or \code{\linkS4class{RasterBrickTS}} object.}

\item{rastermask}{Either a \code{\linkS4class{RasterLayer}} or "compute". Raster layer to use as a mask. When "compute" 
is set raster mask is computed to remove all pixels with incomplete time series.}

\item{nu}{Numeric. Defines the number of EOFs to return. Defaults to return the full set of EOFs.}

\item{gapfill}{Character. Defines the algorithm to be used to interpolate pixels with incomplete temporal profiles. Accepts argument supported as method in function \code{\link[rtsa]{rtsa.gapfill}}.}

\item{centered}{Logical. If TRUE center the input data before EOF computation. Argument is passed through to function \code{\link[sinkr]{eof}}.}

\item{scaled}{Logical. If TRUE scale the input data before EOF computation. Argument is passed through to function \code{\link[sinkr]{eof}}.}

\item{cores}{Integer. Defines the number of CPU to be used for multicore processing. Default to "1" core for singlecore processing. Applies only to the masking step.}

\item{...}{Additional arguments to be passed through to function \code{\link[sinkr]{eof}}.}
}
\value{
Object of class \code{\linkS4class{EOFstack}} containing the following components:
\tabular{rll}{
\tab \code{eof.modes} \tab EOF modes as \code{\linkS4class{RasterBrick}} object\cr
\tab \code{expansion_coefficients} \tab EOF Expansion Coefficients (EC) as \code{\linkS4class{xts}} object\cr
\tab \code{total_variance} \tab Numeric. Total variance of input raster time series\cr
\tab \code{explained_variance} \tab Numeric vector. Percentage of variance explained by each EOF mode with respect to the total variance of input raster time series\cr
\tab \code{center} \tab Center values from each pixel temporal profile as \code{\linkS4class{RasterLayer}} object (only computed if \code{centered = TRUE})\cr
\tab \code{scale} \tab Scale values from each pixel temporal profile as \code{\linkS4class{RasterLayer}} object (only computed if \code{scaled = TRUE})
}
}
\description{
This function conducts an Empirical Orthogonal Function analysis (EOF) via a covariance matrix 
(cov4gappy function) using "sinkr" package especially designed to handle gappy raster time series
}
\details{

}
\examples{
\dontrun{
## create raster time series using the 'pacificSST' data from 'remote' package
require(remote)

data(pacificSST)
pacificSST[which(getValues(pacificSST == 0))] <- NA # set NA values
# create rts object
rasterts <- rts(pacificSST, seq(as.Date('1982-01-15'), as.Date('2010-12-15'), 'months'))

## generate raster mask
raster_mask <- pacificSST[[1]] # create raster mask
names(raster_mask) <- "mask"
values(raster_mask) <- 1 # set raster mask values
raster_mask[which(is.na(getValues(pacificSST[[1]])))] <- 0 # set raster mask values

## compute EOF
# compute the first 10 EOFs
eof_result <- rtsa.eof(rasterts=rasterts, nu=10)
# recursively compute the first 10 EOFs using raster mask
eof_result_recursive <- rtsa.eof(rasterts=rasterts, rastermask=raster_mask, nu=10, recursive=T)
# compute the first 10 EOFs applying centering, scaling 
# and raster mask computing before eof computation
eof_res_masked <- rtsa.eof(rasterts=rasterts, rastermask="compute", nu=10, centered=T, scaled=T)
}

}
\references{
Bjoernsson, H. and Venegas, S.A. (1997). "A manual for EOF and SVD 
analyses of climate data", McGill University, CCGCR Report No. 97-1, 
Montreal, Quebec, 52pp.

Marc, T.H., Losch, M., Wenzel, M., Schroeter, J. (2013). 
On the Sensitivity of Field Reconstruction and Prediction Using 
Empirical Orthogonal Functions Derived from Gappy Data. Journal of Climate, 
26, 9194-9205. \href{http://dx.doi.org/10.6084/m9.figshare.732650}{pdf}
}
\seealso{
\code{\link[sinkr]{eof}}, \code{\link[rtsa]{rtsa.scaleEOF}}, \code{\link[rtsa]{rtsa.gapfill}}
}
\author{
Federico Filipponi
}
\keyword{EOF}
\keyword{PCA}
\keyword{SVD}
\keyword{analysis}
\keyword{series}
\keyword{time}
