% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTSA.R
\name{RTSA}
\alias{RTSA}
\title{RTSA}
\usage{
RTSA(data, outcome = "RR", mc)
}
\arguments{
\item{data}{A data set containing eI, eC, nI, nC or mI, mC, sdI, sdC, nI, nC}

\item{outcome}{Outcome of interst, RR only possibility now}

\item{mc}{minimum clinical relevant outcome}
}
\value{
A TSA object (list for now)
}
\description{
RTSA
}
\examples{
data(perioOxy)
RTSA(data = perioOxy, outcome = "RR", mc = 0.9)
}
