% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTSA.R
\name{RTSA}
\alias{RTSA}
\title{RTSA}
\usage{
RTSA(
  data,
  side,
  outcome = "RR",
  mc,
  alpha = 0.05,
  beta = 0.2,
  futility = "none",
  fixed = FALSE,
  ana_time = NULL,
  method = "IV",
  vartype = "equal",
  sign = NULL,
  fixedStudy = FALSE,
  hksj = FALSE,
  tau.ci.method = "BJ",
  ...
)
}
\arguments{
\item{data}{A data set containing eI, eC, nI, nC or mI, mC, sdI, sdC, nI, nC}

\item{side}{I or II sided hypothesis test}

\item{outcome}{Outcome of interest, RR only possibility now}

\item{mc}{minimum clinical relevant outcome}

\item{alpha}{The type I error}

\item{beta}{The type II error}

\item{futility}{Futility boundaries added to design. Default is "none".}

\item{fixed}{Default is FALSE. Sets heterogeneity to 0 if TRUE.}

\item{method}{Weighting method. Defaults to Mantel-Haenzel (MH).}

\item{vartype}{TBD.}

\item{sign}{TBD.}

\item{fixedStudy}{TBD.}

\item{hksj}{TBD.}

\item{tau.ci.method}{TBD.}

\item{...}{other arguments}
}
\value{
A TSA object (list for now)
}
\description{
RTSA
}
\examples{
data(perioOxy)
RTSA(data = perioOxy, outcome = "RR", mc = 0.7, side = 2, alpha = 0.05, beta = 0.2)
}
