% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTSA.R
\name{RTSA}
\alias{RTSA}
\title{RTSA}
\usage{
RTSA(data, outcome = "RR", mc, alpha = 0.05, beta = 0.2, fixed = FALSE, ...)
}
\arguments{
\item{data}{A data set containing eI, eC, nI, nC or mI, mC, sdI, sdC, nI, nC}

\item{outcome}{Outcome of interest, RR only possibility now}

\item{mc}{minimum clinical relevant outcome}

\item{alpha}{The type I error}

\item{beta}{The type II error}

\item{fixed}{Default is FALSE. Sets heterogeneity to 0 if TRUE.}

\item{...}{other arguments}
}
\value{
A TSA object (list for now)
}
\description{
RTSA
}
\examples{
data(perioOxy)
RTSA(data = perioOxy, outcome = "RR", mc = 0.9)
}
